requirements.txt
Django
django-bootstrap-datepicker-plus
django-bootstrap-form
django-datatables-view
Pillow
pytz
sqlparse
gemmi
django-cleanup
django-crispy-forms
django-robohash-svg
django-translation-flags
django-simple-history
crispy-bootstrap4